pyproject.toml
[project]
name = "pdf-image-extractor"
version = "1.0.0"
description = 'A very simple executable Python package to extract images from a PDF file.'
readme = "README.md" # Optional
requires-python = ">=3.10"
license = { file = "LICENSE.txt" }
keywords = ['python', 'PDF', 'image', 'extraction']
authors = [
    { name = "jnickg", email = "me@jnickg.net" },
]
maintainers = [
    { name = "jnickg", email = "me@jnickg.net" },
]

# For a list of valid classifiers, see https://pypi.org/classifiers/
classifiers = [ # Optional
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
    "fitz",
    "Pillow"
]

[project.urls]
"Homepage" = "https://github.com/jnickg/pdf_image_extractor"
"Bug Reports" = "https://github.com/jnickg/pdf_image_extractor/issues"
"Source" = "https://github.com/jnickg/pdf_image_extractor"

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"
